refactor(Constants): 상수 클래스 속 상수 일부를 사용처 클래스로 이동
