+++ README.md
@@ -146,7 +146,8 @@
 - [x] 클래스 변수 2개 이하로 만들기
 - [x] 자판기 클래스 생성
 - [x] Utils 패키지 생성 및 클래스 이동
-- [x] 컨버터 클래스로 인풋 스트링 처리 기능 이동
+- [x] 컨버터 클래스로 인풋 스트링 처리 기능 분리
+- [x] 상수 클래스 속 상수 일부를 사용처 클래스로 이동
 - [ ] MVC패턴 적용해 분할하기
 - [ ] 메소드가 한가지 일만 하도록 하기
 - [ ] InputController 속 메소드 반복/반복문 줄이고 구조화 하기
